pyproject.toml
[build-system]
requires = [
    "setuptools>=61",
    "wheel",
    "setuptools-scm[toml]>=6.2",
]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "AWS/_version.py"

[project]
name = "robotframework-aws"
description = "Robot Framework library for AWS"
readme = "README.md"
requires-python = ">=3.8"
keywords = ["AWS", "testing"]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dynamic = ["version"]
dependencies = [
  "robotframework >= 6.0.0, < 8.0.0",
  "python-dotenv >= 1.0.0, < 1.1.0",
  "dotmap >= 1.3.30, < 1.4.0",
  "boto3 >= 1.35.7, < 1.36.0",
  "botocore >= 1.35.7, < 1.36.0",
]
